slight refacto of the agent
